@@ -8,32 +8,51 @@
 
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.12'
-gem 'supplejack_client', git: 'git@github.com:DigitalNZ/supplejack_client.git'
+if RUBY_VERSION =~ /2/
+  group :development do
+    gem 'better_errors'
+    gem 'pry-rails'
+  end
 
+  group :development, :test do
+    gem 'pry-byebug'
+  end
+end
+
+gem 'rails', '~> 4.1.0'
+gem 'supplejack_client', git: 'https://github.com/DigitalNZ/supplejack_client.git'
+
+gem 'mysql2', '~> 0.3.18'
 gem 'json'
 gem 'will_paginate'
 gem 'jquery-rails'
+gem 'activeresource'
+gem 'devise'
+gem "codeclimate-test-reporter", group: :test, require: nil
+gem 'haml', '~> 4.0.6'
+gem 'rubocop', require: false
 
 group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
+  gem 'sass-rails',   '~> 4.0.3'
+  gem 'coffee-rails', '~> 4.0.0'
   gem 'uglifier', '>= 1.0.3'
-  gem 'compass-rails'
-  gem 'zurb-foundation', '~> 4.0.0'
+  gem 'foundation-rails', '~> 5.4.5.0'
+  gem 'font-awesome-rails', '~> 4.3.0.0'
 end
 
 group :development do
+  gem "erb2haml"
   gem 'thin'
   gem 'quiet_assets'
-  gem 'better_errors'
   gem 'binding_of_caller'
 end
 
 group :test, :development do
-  gem 'debugger'
   gem "rspec-rails", "2.14"
   gem 'oily_png'
+  gem 'minitest'
+  gem 'simplecov'
+  gem 'ffaker',   '~> 2.1.0'
 end
 
 group :test do
@@ -41,6 +60,7 @@ group :test do
   gem 'factory_girl_rails', "= 1.2.0"
   gem 'spork', '>= 1.0.0rc3'
   gem 'vcr', '2.9.0'
+  gem 'faker'
 end
 
 group :production do
