@@ -1,13 +1,15 @@
-# The majority of The Supplejack Website code is Crown copyright (C) 2014, New Zealand Government,
-# and is licensed under the GNU General Public License, version 3. Some components are 
-# third party components that are licensed under the MIT license or other terms. 
-# See https://github.com/DigitalNZ/supplejack_website for details. 
-# 
-# Supplejack was created by DigitalNZ at the National Library of NZ and the Department of Internal Affairs. 
+# The majority of The Supplejack Website code is Crown copyright (C) 2014,
+# New Zealand Government, and is licensed under the GNU General Public License,
+# version 3. Some components are third party components that are licensed under
+# the MIT license or other terms.
+# See https://github.com/DigitalNZ/supplejack_website for details.
+#
+# Supplejack was created by DigitalNZ at the National Library of NZ and
+# the Department of Internal Affairs.
 # http://digitalnz.org/supplejack
 
+# StaticPage controller deals with the default info pages
 class StaticPagesController < ApplicationController
-
   def become_a_partner
   end
 
@@ -22,5 +24,4 @@ class StaticPagesController < ApplicationController
 
   def base_template
   end
-
 end
