@@ -1,33 +1,39 @@
-# The majority of The Supplejack Website code is Crown copyright (C) 2014, New Zealand Government,
-# and is licensed under the GNU General Public License, version 3. Some components are 
-# third party components that are licensed under the MIT license or other terms. 
-# See https://github.com/DigitalNZ/supplejack_website for details. 
-# 
-# Supplejack was created by DigitalNZ at the National Library of NZ and the Department of Internal Affairs. 
+# The majority of The Supplejack Website code is Crown copyright (C) 2014,
+# New Zealand Government, and is licensed under the GNU General Public License,
+# version 3. Some components are third party components that are licensed under
+# the MIT license or other terms.
+# See https://github.com/DigitalNZ/supplejack_website for details.
+#
+# Supplejack was created by DigitalNZ at the National Library of NZ and
+# the Department of Internal Affairs.
 # http://digitalnz.org/supplejack
 
+# SearchTab
 class SearchTab
-
+  TABS = %w(All Images Audio Videos Sets)
   attr_reader :tab
-  
-  def initialize(tab='all')
-    @tab = tab.present? ? tab : 'all'
+
+  def initialize(tab = 'All')
+    @tab = tab.present? ? tab : 'All'
   end
 
-  def self.add_type_facets(search, type)
-    if valid_type_facets.include?(type)
+  def self.add_category_facets(search, category)
+    if valid_category_facets.include?(category)
       search.and ||= {}
-      search.and[:type] ||= {}
-      search.and[:type] = type
+      search.and[:category] ||= {}
+      search.and[:category] = category
     end
   end
 
-  def self.valid_type_facets
-    Search.new.facet_values('type').keys
+  def self.valid_category_facets
+    Search.new.facet_values('category').keys
   end
 
   def self.sorted_counts(counts)
-    sorted = Hash[counts.sort_by{|k, v| v}.reverse[0..4]]
+    sorted = {}
+    TABS.each do |tab|
+      sorted[tab] = counts[tab]
+    end
     sorted
   end
 
@@ -36,11 +42,10 @@ class SearchTab
   end
 
   def all?
-    tab == 'all'
+    tab == 'All'
   end
-  
+
   def value
     @tab if tab.present?
   end
-
 end
