@@ -8,12 +8,13 @@
 
 require File.expand_path('../boot', __FILE__)
 
-# require 'rails/all'
+require 'rails/all'
 require 'action_controller/railtie'
 require 'action_mailer/railtie'
 require 'active_resource/railtie'
 require 'rails/test_unit/railtie'
 require 'sprockets/railtie'
+# require 'active_record'
 
 if defined?(Bundler)
   # If you precompile assets before deploying to production, use this line
