@@ -16,13 +16,14 @@ Demo::Application.configure do
 
   # Log error messages when you accidentally call methods on nil.
   config.whiny_nils = true
+  config.eager_load = false
 
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
 
   # Don't care if the mailer can't send
-  config.action_mailer.raise_delivery_errors = false
+  config.action_mailer.raise_delivery_errors = true
 
   # Print deprecation notices to the Rails logger
   config.active_support.deprecation = :log
@@ -42,4 +43,15 @@ Demo::Application.configure do
 
   # Expands the lines which load the assets
   config.assets.debug = true
+
+  config.action_mailer.delivery_method = :smtp
+
+  config.action_mailer.smtp_settings = {
+    :address  => "localhost",
+    :domain  => "beta.dnz0a.digitalnz.org/",
+    :port  => 25
+  }
+
+  # For Devise
+  config.action_mailer.default_url_options = { host: 'localhost', port: 3000 }
 end
