@@ -1,9 +1,9 @@
 # The majority of The Supplejack Website code is Crown copyright (C) 2014, New Zealand Government,
-# and is licensed under the GNU General Public License, version 3. Some components are 
-# third party components that are licensed under the MIT license or other terms. 
-# See https://github.com/DigitalNZ/supplejack_website for details. 
-# 
-# Supplejack was created by DigitalNZ at the National Library of NZ and the Department of Internal Affairs. 
+# and is licensed under the GNU General Public License, version 3. Some components are
+# third party components that are licensed under the MIT license or other terms.
+# See https://github.com/DigitalNZ/supplejack_website for details.
+#
+# Supplejack was created by DigitalNZ at the National Library of NZ and the Department of Internal Affairs.
 # http://digitalnz.org/supplejack
 
 Supplejack.configure do |config|
@@ -27,10 +27,11 @@ Supplejack.configure do |config|
   # When you ask for the facets, they are going to be ordered in the
   # order presented here
   config.facets = [
-    :source_provider_name,
-    :source_contributor_name,
-    :source_website_name,
-    :type
+    :category,
+    :content_partner,
+    :primary_collection,
+    :usage,
+    :date
   ]
   #
   # ===> Facet values sorting
@@ -45,14 +46,22 @@ Supplejack.configure do |config|
   # record. :default will return the default set of fields.
   #
   config.fields = [
-    :all,
-    :type
+    :default,
+    :source_contributor_name,
+    :display_collection,
+    :display_content_partner,
+    :source_url,
+    :thumbnail_url,
+    :display_date,
+    :creator,
+    :category,
+    :subject
   ]
 
   # ===> Number of facet values
   # This will limit the number of facet values returned for each facet
   # Be carefull not to make it too high for performance reasons
-  # config.facets_per_page = 10
+  config.facets_per_page = 20
   #
   # ===> Per Page
   # Number of results returned per page
@@ -80,7 +89,7 @@ Supplejack.configure do |config|
   #   search.type
   #
   config.search_attributes = [
-    :type
+    :subject
   ]
   #
   # ===> Record klass
@@ -91,7 +100,7 @@ Supplejack.configure do |config|
   # config.record_klass = "Record"
   #
   # ===> Enable Debugging
-  # Set this flag to true in order to get display errors and the actual SOLR requests 
+  # Set this flag to true in order to get display errors and the actual SOLR requests
   # in the logs generated by the API.
   #
   config.enable_debugging = true
@@ -99,5 +108,5 @@ Supplejack.configure do |config|
   # ===> Enable Caching
   # Set this flag to true in order to cache the facet_value response and the search counts
   #
-  config.enable_caching = true  
+  config.enable_caching = false
 end
