@@ -6,6 +6,9 @@
 # Supplejack was created by DigitalNZ at the National Library of NZ and the Department of Internal Affairs. 
 # http://digitalnz.org/supplejack
 
+require 'simplecov'
+SimpleCov.start
+
 require 'rubygems'
 require 'spork'
 #uncomment the following line to use spork with the debugger
@@ -22,7 +25,9 @@ Spork.prefork do
   require 'rspec/rails'
   require 'webmock/rspec'
   # require 'rspec/autorun'
-
+  
+  require "codeclimate-test-reporter"
+  CodeClimate::TestReporter.start
   # Requires supporting ruby files with custom matchers and macros, etc,
   # in spec/support/ and its subdirectories.
   Dir[Rails.root.join("spec/support/**/*.rb")].each { |f| require f }
@@ -54,6 +59,8 @@ Spork.prefork do
     # the seed, which is printed after each run.
     #     --seed 1234
     config.order = 'random'
+    config.include Devise::TestHelpers, :type => :controller
+    config.extend ControllerMacros, :type => :controller
   end
 end
 
